Add physical address as an option in follow/contact section
If the theme is used by someone retro enough to post the address, then
this option provides generalized way to put a standard html address on
the page. It might be an useful option in case someone wants to use it,
but it is easily disabled by just commenting out the section.